% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adi.R
\name{adi}
\alias{adi}
\title{Alternating Direction Implicit Method for Separable-plus-banded Covariance}
\usage{
adi(
  A1,
  A2,
  band,
  y,
  theta = 1e-05,
  rho = NULL,
  adapt = TRUE,
  maxiter = 200,
  tol = 10^-9
)
}
\arguments{
\item{A1}{temporal kernel of the separable-plus-banded model}

\item{A2}{spatial kernel of the separable-plus-banded model}

\item{band}{symbol of the banded part of the separable-plus-banded model}

\item{y}{right-hand side vector}

\item{theta}{regularization}

\item{rho}{initial value of the shift parameter, defaults to the square-root
of the smaller one of the condition numbers of \code{A1} and
\code{A2}. It should should be kept as default unless A1 or A2 is
(exactly) low-rank.}

\item{adapt}{whether the shift should change in between iterations}

\item{maxiter}{maximum number of iterations}

\item{tol}{relative tolerance - stopping criterion}
}
\value{
list of 3 elements: the solution as a vector, the number of ADI
iterations and the vector of PCG iterations (one number per one ADI
iteration)
}
\description{
Calculates the solution to the inverse problem involving the
separable-plus-banded structure as the left-hand side, where the banded part
is stationary.
}
\examples{
K <- 30
A1 <- brownian_cov(K)
A1 <- A2 <- A1/sum(diag(A1))
B <- matrix(c(4,2,2,1)/4,2)/K^2 # to have trace 1
theta <- 1e-5
eigvals <- Re(stats::fft(to_book_format(B,K,K)))
B <- Re(stats::fft(eigvals,inverse = TRUE)/(4*K^2))
B <- B[1:K,1:K]
x <- runif(K^2)
X <- matrix(x,ncol=K)
y <- c(A1 \%*\% X \%*\% A2) + BXfast(eigvals,x) + theta*x
x_adi <- adi(A1, A2, B, y, theta, adapt=TRUE,maxiter=100,tol=10^-7)
sum((x-x_adi$x)^2)/sum(x^2)
}
