% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adi.R
\name{adi_diag}
\alias{adi_diag}
\title{Alternating Direction Implicit Method for Separable-plus-diagonal Covariance}
\usage{
adi_diag(
  A1,
  A2,
  B_diag,
  y,
  theta = 0,
  rho = NULL,
  adapt = F,
  maxiter = 200,
  tol = 10^-6
)
}
\arguments{
\item{A1}{temporal kernel of the separable-plus-diagonal model}

\item{A2}{spatial kernel of the separable-plus-diagonal model}

\item{B_diag}{the diagonal of the model}

\item{y}{right-hand side vector}

\item{theta}{regularization}

\item{rho}{initial value of the shift parameter, defaults to the square-root
of the smaller one of the condition numbers of \code{A1} and
\code{A2}. It should should be kept as default unless A1 or A2 is
(exactly) low-rank.}

\item{adapt}{whether the shift should change in between iterations}

\item{maxiter}{maximum number of iterations}

\item{tol}{relative tolerance - stopping criterion}
}
\value{
list of 3 elements: the solution as a vector, the number of ADI
iterations and the vector of PCG iterations (one number per one ADI
iteration)
}
\description{
Calculates the solution to the inverse problem involving the
separable-plus-diagonal structure as the left-hand side.
}
\examples{
K <- 30
A1 <- brownian_cov(K)
A1 <- A2 <- A1/sum(diag(A1))
B_diag <- array(runif(K^2),c(K,K))
B_diag <- B_diag/sum(B_diag)
theta <- 1e-5
x <- runif(K^2)
X <- matrix(x,ncol=K)
y <- c(A1 \%*\% X \%*\% A2) + c(B_diag*X) + theta*x
x_adi <- adi_diag(A1, A2, B_diag, y, theta, adapt=TRUE,maxiter=100,tol=10^-7)
sum((x-x_adi$x)^2)/sum(x^2)
}
