% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg.R
\name{apply_lhs}
\alias{apply_lhs}
\title{Fast Application of R-separable Covariance}
\usage{
apply_lhs(Est, Y)
}
\arguments{
\item{Est}{list of 3, as produced by \code{\link{scd_est}}:
\code{A} - array of size \code{R} x \code{K1} x \code{K1},
the first \code{R} temporal kernels;
\code{B} - array of size \code{R} x \code{K2} x \code{K2},
the first \code{R} spatial kernels;
\code{sigma} - vector of length \code{R},
the separable component scores}

\item{Y}{matrix argument}
}
\value{
`Est(Y)'
}
\description{
Calculates 'Est(Y)' = \code{sigma1 A1 Y B1 + ... + sigmaR AR Y BR}.
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
Y <- array(runif(3*4),c(3,4))
Est <- scd_est(X,R=2)
apply_lhs(Est, Y)
}
