% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvband_nonstat.R
\name{banded_nonstat_est}
\alias{banded_nonstat_est}
\title{Estimate a Non-stationary Banded Covariance}
\usage{
banded_nonstat_est(X, d, A1 = NULL, A2 = NULL)
}
\arguments{
\item{X}{data set, array of size \code{N} x \code{K1} x \code{K2}}

\item{d}{integer >1, the discrete bandwidth}

\item{A1}{matrix of size \code{K1} x \code{K1}, the temporal kernel, defaults
to zero}

\item{A2}{matrix of size \code{K2} x \code{K2}, the spatial kernel, defaults
to zero}
}
\value{
array B of size \code{d} x \code{2*d-1} x \code{K1*K2}
specifying the banded part of the covariance in the following format:
\code{B[1+i,1+j,]} is a vector giving the off-diagonal of C
corresponding
to shift i in time and shift j in space, while \code{B[1+i,2*d-j,]} is
a vector giving the off-diagonal of C corresponding to shift i in
time and negative shift j in space, suitably adjunct by zeros.
Examples:
\code{B[1,1,]} gives the diagonal of C;
\code{B[3,2,]} gives covariances between locations on the
domain separated by 2 in time and 1 in space in the same direction
(including e.g. \code{cov(X[1,1], X[3,2])} or
\code{cov(X[2,4], X[4,5])});
\code{B[3,2*d-1,]} gives covariances between locations on the domain
separated by 2 in time and 1 in space but in the opposite directions
(including e.g. \code{cov(X[3,1], X[1,2])}).
}
\description{
Calculates the banded version of the empirical covariance for the purpose
of cross-validating bandwidth without assuming stationarity. This can be done
simply by 1) calculating the empirical covariance and 2) setting entries
outside of a given band to zero, which is however computationally wasteful,
because many calculations from step 1 are thrown away in step 2.
Instead, here we calculate directly only those entries of the empirical
covariance, which will be kept non-zero.
}
\examples{
X <- array(runif(20*3*4), c(20,3,4))
B <- banded_nonstat_est(X,2)
B[1,1,] # diagonal of the empirical covariance
}
