% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvband_nonstat.R
\name{banded_nonstat_ip}
\alias{banded_nonstat_ip}
\title{Inner Product of an Observation with itself through a Non-stationary Banded
Covariace}
\usage{
banded_nonstat_ip(X, B)
}
\arguments{
\item{X}{observation, a \code{K1} x \code{K2} matrix}

\item{B}{banded non-stationary covariance in the form of output of
\code{banded_nonstat_est()}}
}
\value{
the inner product, i.e. a number
}
\description{
Inner Product of an Observation with itself through a Non-stationary Banded
Covariace
}
\examples{
X <- array(runif(20*3*4), c(20,3,4))
B <- banded_nonstat_est(X,2)
banded_nonstat_ip(X[1,,],B)
}
