% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvR.R
\name{blup}
\alias{blup}
\title{Best Linear Unbiased Prediction in R-separable Model}
\usage{
blup(Xnew, Res, I, J, eps)
}
\arguments{
\item{Xnew}{new observation to be predicted, given as a matrix}

\item{Res}{list of 3, R-separable covariance estimated by \code{scd_est()}}

\item{I}{indices of unobserved rows}

\item{J}{indices of unobserved columns}

\item{eps}{ridge regularization constant}
}
\value{
prediction o \code{Xnew} given as a matrix
}
\description{
Best Linear Unbiased Prediction in R-separable Model
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
Res <- scd_est(X,2)
Xnew <- array(runif(3*4), c(3,4))
blup(Xnew, Res, 3, 4, 1e-5)
}
