% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brownian_cov}
\alias{brownian_cov}
\title{Covariance of the Wiener Process}
\usage{
brownian_cov(K)
}
\arguments{
\item{K}{grid size}
}
\value{
the covariance, i.e. matrix of size \code{K} x \code{K}
}
\description{
Covariance of the Wiener Process
}
\examples{
brownian_cov(20)
}
