% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BXfast}
\alias{BXfast}
\title{Fast Multiplication for a Two-level Toeplitz Matrix}
\usage{
BXfast(eigvals, x)
}
\arguments{
\item{eigvals}{eigenvalues of a Two-level Toeplitz Matrix}

\item{x}{vectorization of matrix observation}
}
\value{
vector vec(Y), which can be matrized to obtain the product Y
}
\description{
We wish to calculate the tensor-matrix product Y = B X fast, where B is a
stationary tensor, i.e. vec(Y) = mat(B) vec(X) is a product of a Two-level
Toeplitz matrix with a vector, which can be calculated fast using 2D FFT.
}
\examples{
band <- matrix(4:1,2)
eigvals <- stats::fft(to_book_format(band,3,3))
BXfast(Re(eigvals), runif(9))
}
