% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adi.R
\name{calc_prec}
\alias{calc_prec}
\title{Calculates Eigenvalues of the Preconditioner}
\usage{
calc_prec(M)
}
\arguments{
\item{M}{the circulant symbol matrix generated by \code{to_book_format()}}
}
\value{
matrix of eigenvalues, one block of the preconditioner corresponds to
one column of the eigenvalue matrix
}
\description{
Computes eigenvalues of the blocks of the circulant preconditioner, which is
denoted as \code{c^{(1)}} in Chan, R.H.F., & Jin, X.Q. (2007) An
introduction to iterative Toeplitz solvers.
}
\examples{
eigvals <- matrix(9:1,3)
M <- to_book_format(eigvals,3,3)
calc_prec(M)
}
