% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvband.R
\name{cvband_pred}
\alias{cvband_pred}
\title{Cross-validation for Bandwidth Selection in Separable-plus-banded Model via
Prediction}
\usage{
cvband_pred(X, Folds = 10, maxd = 20, mind = 1, perc = c(1/4, 1/4))
}
\arguments{
\item{X}{data, array of size N x K1 x K2}

\item{Folds}{number of folds}

\item{maxd}{maximum bandwidth value to check}

\item{mind}{minimum bandwidth value to check}

\item{perc}{vector of 2 values in (0,1) giving the ratio of rows and columns
to be predicted}
}
\value{
vector of cross-validated prediction errors for
d=\code{mind},...,\code{maxd}
}
\description{
Examines the prediction performance of the separable-plus-banded model for
different (discrete) bandwidths d=\code{mind},...,\code{maxd} via a
cross-validated scheme: for every surface in the current hold-out sample,
a part of the surface (a fixed number of last few rows and columns
give by \code{perc}) is predicted based on the remainder of the given surface,
and relative prediction error is calculated.
}
\examples{
N <- 4
K1 <- 4
K2 <- 4
set.seed(517)
X <- array(0,c(N,K1,K2))
A <- matrix(rnorm(K1^2),K1)
B <- matrix(rnorm(K2^2),K2)
A <- A \%*\% t(A)
B <- B \%*\% t(B)
A <- mat_root(A)
B <- mat_root(B)
for(n in 1:N){
  X[n,,] <- A \%*\% X[n,,] \%*\% B + matrix(rnorm(K1*K2),K1)
}
( cvscores <- cvband_pred(X,2,2,0) )
min(localMaxima(-cvscores))
}
