% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvR.R
\name{cvR}
\alias{cvR}
\title{Cross-validation for Degree-of-Separability Selection}
\usage{
cvR(X, Folds = 10, maxR = 7, maxiter = 10)
}
\arguments{
\item{X}{data set, array of size \code{N} x \code{K1} x \code{K2}}

\item{Folds}{number of folds}

\item{maxR}{maximum degree-of-separability considered
(defaults to log-sample size)}

\item{maxiter}{maximum number of iterations for \code{scd_est()}}
}
\value{
CV objective
}
\description{
Calculates the objective value of the cross-validation criterion to select
the degree of separability R=1,...,maxR best fitting the data.
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
cvscores <- cvR(X,10,5)
min(localMaxima(-cvscores)) # best R is chosen as the smallest local minimum
                            # of the CV objective
}
