% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvR.R
\name{cvR_pred}
\alias{cvR_pred}
\title{Cross-validation for Degree-of-separability Selection via Prediction}
\usage{
cvR_pred(X, Folds = 10, maxR = 7, maxiter = 10, perc = NULL)
}
\arguments{
\item{X}{data set, array of size \code{N} x \code{K1} x \code{K2}}

\item{Folds}{number of folds}

\item{maxR}{maximum degree-of-separability considered}

\item{maxiter}{maximum number of iterations for scd_est()}

\item{perc}{number in (0,1) giving the portion of rows and columns held out
for prediction, defaults to 1/2, in that case 3/4 of every
held-out surface is predicted based on the remaining 1/4}
}
\value{
vector of prediction performances for different values of R
}
\description{
Examines the prediction performance of R-separable model for
the degree of separability R=1,...,maxR via a cross-validated scheme:
for every surface in the current hold-out sample, part of the surface
(a fixed number of rows and columns given by \code{perc}) is predicted based
on the remainder, and relative prediction error is calculated. Two random
splits of every observation are predicted.
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
cvscores <- cvR_pred(X,10,5)
min(localMaxima(-cvscores)) # best R is chosen as the smallest local minimum
                            # of the CV objective
}
