% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spb_goftest.R
\name{dis2sep}
\alias{dis2sep}
\title{Projected distances to separability, i.e. the bootstrapped statistic.}
\usage{
dis2sep(X, d, mu, I, J, stationary)
}
\arguments{
\item{X}{data array of size \code{N} x \code{K1} x \code{K2}}

\item{d}{the bandwidth, an integer between \code{0} and \code{min(K1,K2)}}

\item{mu}{matrix of size \code{K1} x \code{K2}, the mean, defaults to the
empirical mean}

\item{I}{number of temporal eigenfunctions used to assess the fit}

\item{J}{number of spatial eigenfunctions used to assess the fit}

\item{stationary}{a flag whether the banded part should be stationary,
defaults to \code{TRUE}}
}
\value{
array of projected distances to separability
}
\description{
Projected distances to separability, i.e. the bootstrapped statistic.
}
