% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvR.R
\name{eigvals}
\alias{eigvals}
\title{Maximum and Minimum Eigenvalue of an R-separable Covariance}
\usage{
eigvals(Est, tol = 1e-10, maxiter = 10000)
}
\arguments{
\item{Est}{R-separable covariance given as a list of 3:
\code{A} - array of size R x K1 x K1, the first R temporal kernels;
\code{B} - array of size R x K2 x K2, the first R spatial kernels;
\code{sigma} - vector of length R, the separable component scores}

\item{tol}{stopping criterion}

\item{maxiter}{maximum number of iterations}
}
\value{
list of 2: \code{min} - the minimum eigenvalue;
\code{max} - the maximum eigenvalue
}
\description{
Maximum and Minimum Eigenvalue of an R-separable Covariance
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
Res <- scd_est(X,2)
eigvals(Res)
}
