% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epanechnik}
\alias{epanechnik}
\title{Epanechnikov Kernel}
\usage{
epanechnik(x)
}
\arguments{
\item{x}{double}
}
\value{
value of epanechnikov kernel at \code{x}
}
\description{
Returns f(x) where f is the Epanechnikov kernel supported on (-1,1).
}
\examples{
epanechnik(0)
}
