% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ext2arr}
\alias{ext2arr}
\title{Extend to Array}
\usage{
ext2arr(M)
}
\arguments{
\item{M}{a numeric matrix}
}
\value{
array of size \code{1} x \code{dim(M)[1]} x \code{dim(M)[2]}
}
\description{
Extends a matrix M into array of size 1 x \code{dim(M)[1]} x
\code{dim(M)[2]}.
}
\examples{
ext2arr(matrix(rnorm(9),3))
}
