% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{frobenius}
\alias{frobenius}
\title{Frobenius Norm of an Array}
\usage{
frobenius(X)
}
\arguments{
\item{X}{array of arbitrary dimension and size}
}
\value{
a number, the Frobenius norm of \code{X}
}
\description{
Frobenius Norm of an Array
}
