% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{frobenius_piece}
\alias{frobenius_piece}
\title{Frobenius Norm of a rank-3 Matrix}
\usage{
frobenius_piece(a1, b1, a2, b2, beta)
}
\arguments{
\item{a1}{left singular vector of the minuend}

\item{b1}{right singular vector of the minuend}

\item{a2}{left singular vector of the subtrahend}

\item{b2}{left singular vector of the minuend}

\item{beta}{corresponding to the constant matrix 1 otimes 1}
}
\value{
a number, the Frobenius norm
}
\description{
Calculates the Frobenius norm of a single sub-matrix of the difference of the
Van Loan's permutations of a separable-plus-banded covariance, which must be
at most rank 3, with left factors \code{a1}, \code{b1} and \code{1}, and
right factors \code{a2}, \code{b2}, \code{1}. Gram-Schmidt orthogonalization
can be used to calculate the Frobenius norm of such a matrix of size \code{K}
x \code{K} in only 3*\code{K} operations, where \code{K} is the dimension of
the matrix.
}
\examples{
a1 <- runif(5)
b1 <- runif(10)
a2 <- runif(5)
b2 <- runif(10)
beta <- runif(1)
frobenius_piece(a1,b1,a2,b2,beta)
}
