% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{frobenius_piece_slow}
\alias{frobenius_piece_slow}
\title{Frobenius Norm of a rank-3 Matrix (Vanilla)}
\usage{
frobenius_piece_slow(a1, b1, a2, b2, beta)
}
\arguments{
\item{a1}{left singular vector of the minuend}

\item{b1}{right singular vector of the minuend}

\item{a2}{left singular vector of the subtrahend}

\item{b2}{left singular vector of the minuend}

\item{beta}{corresponding to the constant matrix \code{1} otimes \code{1}}
}
\value{
a number, the Frobenius norm
}
\description{
Calculates the Frobenius norm of a single sub-matrix of the difference of the
Van Loan's permutations of a separable-plus-banded covariance, which must be
at most rank 3, with left factors \code{a1}, \code{b1} and \code{1}, and
right factors \code{a2}, \code{b2}, 1. This is a vanilla version used for
matrices with \code{ncol <= 3} or \code{nrow <= 3}, because the Gram-Schmidt
orthogonalization cannot be used for such a small matrix.
}
\examples{
a1 <- runif(5)
b1 <- runif(10)
a2 <- runif(5)
b2 <- runif(10)
beta <- runif(1)
frobenius_piece(a1,b1,a2,b2,beta)
}
