% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{frobenius_spb}
\alias{frobenius_spb}
\title{Frobenius Norm of a Difference of Two Separable-plus-banded Covariances}
\usage{
frobenius_spb(Chat, C = NULL, relative = T)
}
\arguments{
\item{Chat}{separable-plus-banded model \code{A1} x \code{A2} + \code{B},
i.e. a list of 3 elements:
matrix \code{A1}, matrix \code{A2}, the symbol of \code{B}}

\item{C}{as \code{Chat} above}

\item{relative}{flag whether the error relative to \code{C} should be computed,
defaults to \code{TRUE}}
}
\value{
a number, the Frobenius norm
}
\description{
Frobenius Norm of a Difference of Two Separable-plus-banded Covariances
}
\examples{
X <- array(runif(20*4*5), c(20,4,5))
Y <- array(runif(20*4*5), c(20,4,5))
Chat <- spb(X,2)
C <- spb(Y,2)
frobenius_spb(Chat,C)
}
