% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{frobenius_spb_nonstat}
\alias{frobenius_spb_nonstat}
\title{Frobenius Norm of a Separable-plus-banded Covariance with Non-stationary
Banded Part}
\usage{
frobenius_spb_nonstat(A1, A2, B)
}
\arguments{
\item{A1}{matrix, the temporal kernel}

\item{A2}{matrix, the spatial kernel}

\item{B}{3-dimensional array specifying the banded part in the form of an
output of \code{\link{banded_nonstat_est}}}
}
\value{
the value of the Frobenius norm
}
\description{
Frobenius Norm of a Separable-plus-banded Covariance with Non-stationary
Banded Part
}
\examples{
N <- 20
K1 <- 10
K2 <- 10
delta <- 2
X <- array(runif(N*K1*K2),c(N,K1,K2))
Est <- spt(X,delta)
B <- banded_nonstat_est(X,delta,Est$A1,Est$A2)
frobenius_spb_nonstat(Est$A1,Est$A2,B)
}
