% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{localMaxima}
\alias{localMaxima}
\title{Find Local Maxima of a Vector}
\usage{
localMaxima(x)
}
\arguments{
\item{x}{a vector}
}
\value{
maximum
}
\description{
Find Local Maxima of a Vector
}
\examples{
localMaxima(runif(9))
}
