% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_root}
\alias{mat_root}
\title{Matrix Square-root}
\usage{
mat_root(X)
}
\arguments{
\item{X}{a symmetric matrix}
}
\value{
a symmetric matrix \code{A}
}
\description{
For a symmetric matrix \code{X} calculates the matrix \code{A} such that
\code{X} = \code{A}^2.
}
\examples{
X <- matrix(runif(9),3)
X <- X \%*\% t(X)
mat_root(X)
}
