% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg.R
\name{pcg}
\alias{pcg}
\title{Preconditioned Conjugate Gradient for R-separable Covariance}
\usage{
pcg(Est, Y, eps, Ig = NULL, tol = 1e-10)
}
\arguments{
\item{Est}{list of 3 giving the left-hand side, produced by
\code{\link{scd_est}}}

\item{Y}{the right-hand side matrix}

\item{eps}{ridge regularization constant}

\item{Ig}{the initial guess, random by default}

\item{tol}{the tolerance used for stopping}
}
\value{
list of 2: \code{U} - the solution X as a matrix;
\code{iter} - the number of iterations
}
\description{
PCG method for solving \code{A1 X B1 + ... + AR X BR + eps X = Y}
using the preconditioner \code{(A1 x B1 + eps I)^(-1)}
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
X_new <- array(runif(3*4),c(3,4))
Est <- scd_est(X,R=2)
Y <- apply_lhs(Est, X_new)
X_hat <- pcg(Est, Y, 1e-5)
norm(X_new - X_hat$U, type="F")/norm(X_new, type="F")
}
