% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adi.R
\name{pcgtt}
\alias{pcgtt}
\title{Preconditioned Conjugate Gradient for Two-level Toeplitz Left-hand Side}
\usage{
pcgtt(M, b, init, eigvals_prec, tol = 1e-07)
}
\arguments{
\item{M}{symbol of the circulant embedding of a two-level Toeplitz matrix,
generated from the symbol the two-level Toeplitz matrix by
\code{to_book_format()}}

\item{b}{right-hand side vector}

\item{init}{initial guess for solution}

\item{eigvals_prec}{eigenvalues of the circulant preconditioner generated by
\code{calc_prec()}}

\item{tol}{relative tolerance - stopping criterion}
}
\value{
the solution as a vector
}
\description{
Solves the second ADI equation, i.e. a linear system with two-level Toeplitz
left-hand side.
}
\examples{
eigvals <- matrix(9:1,3)
x <- stats::runif(9)
M <- to_book_format(eigvals,3,3)
eigvals_prec <- calc_prec(M)
pcgtt(M, x, runif(9), eigvals_prec)
}
