% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scd.R
\name{scd}
\alias{scd}
\title{Separable Component Decomposition of the Empirical Covariance}
\usage{
scd(
  X,
  R = NULL,
  B = NULL,
  mu = NULL,
  predict = FALSE,
  maxiter = 10,
  maxR = NULL,
  Folds = NULL,
  perc = c(1/4, 1/4)
)
}
\arguments{
\item{X}{data array of size \code{N} x \code{K1} x \code{K2}}

\item{R}{integer, the degree-of-separability}

\item{B}{array of size \code{R} x \code{K2} x \code{K2} specifying the
initial values for \code{B1},...,\code{BR}. Defaults to partially
traced initializations. If \code{R} not provided by the user,
\code{B} should be array of size \code{K2} x \code{K2}, and this
single starting point is used for every degree-of-separability.}

\item{mu}{matrix of size \code{K1} x \code{K2}, the mean, defaults to the
empirical mean}

\item{predict}{flag whether prediction-based cross-validation should be used}

\item{maxiter}{integer, maximum number of power iterations, defaults to 10}

\item{maxR}{integer, maximum value of the degree-of-separability \code{R}
to be probed via CV (used only when \code{R} not provided
directly)}

\item{Folds}{number of folds for cross-validation, defaults to 10}

\item{perc}{vector of 2, if \code{predict=T}, what proportions (in time and
space) of every observation should be predicted. When
\code{perc}>=1, specific number of grid lines is predicted, e.g.
\code{perc}=1 corresponds to one-step-ahead prediction (in both
space and time)}
}
\value{
list of the following objects:list of 3:
\itemize{
\item \code{A} - array of size \code{R} x \code{K1} x \code{K1}, the first
\code{R} temporal kernels;
\item \code{B} - array of size \code{R} x \code{K2} x \code{K2}, the first
\code{R} spatial kernels;
\item \code{sigma} - vector of length \code{R}, the separable component scores
}

if cross-validation was conducted, the list contains also:
\itemize{
\item \code{cv} - the CV objective values, a smaller value is better
\item \code{R} - the chosen bandwidth
}
}
\description{
Performs the power iteration method to find the separable expansion for
\code{C = cov(X)}, i.e. to find the decomposition \code{C = sigma1 A1} x
\code{B1} + \code{sigma2 A2} x \code{B2} + ... + \code{sigmaR AR} x
\code{BR}. When the degree-of-separability \code{R} is not provided,
cross-validation (CV) is performed.
}
\examples{
N <- 10
K1 <- K2 <- 10
x <- 1:K1/(K1+1)
e0 <- rep(1,K1)
e1 <- sin(2*pi*x)
e2 <- cos(2*pi*x)
X <- array(0,c(N,K1,K2))
set.seed(123)
for(n in 1:N){
  X[n,,] <- 2*rnorm(1)*outer(e0,e0) + rnorm(1)*outer(e1,e1) +
    rnorm(1)*outer(e2,e2)
}
SCD <- scd(X)
SCD$R # the degree-of-separability chosen by fit-based CV
SCD <- scd(X,predict=TRUE)
SCD$R # the degree-of-separability leading to best predictions
SCD <- scd(X,predict=TRUE,perc=1)
SCD$R # the degree-of-separability leading to best one-step-ahead (both in
      # space and time) predictions
}
