% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIP.R
\name{scd_est}
\alias{scd_est}
\title{Estimate the Separable Component Decomposition}
\usage{
scd_est(X, R = 1, maxiter = 10, B = NULL)
}
\arguments{
\item{X}{data set, array of size \code{N} x \code{K1} x \code{K2}}

\item{R}{integer, the degree of separability}

\item{maxiter}{maximum number of iteration}

\item{B}{a spatial kernel used as a starting point}
}
\value{
list of 3:
\code{A} - array of size \code{R} x \code{K1} x \code{K1},
the first \code{R} temporal kernels;
\code{B} - array of size \code{R} x \code{K2} x \code{K2},
the first \code{R} spatial kernels;
\code{sigma} - vector of length \code{R},
the separable component scores
}
\description{
Calculates the R-separable estimator of the covariance of X via the
generalized power iteration, starting from B.
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
scd_est(X,R=2)
}
