% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvband.R
\name{seppband_psd}
\alias{seppband_psd}
\title{Positivize a Separable-plus-banded Estimator}
\usage{
seppband_psd(Res)
}
\arguments{
\item{Res}{list of 3 specifying the separable-plus-banded estimator:
A1 - the temporal kernel;
A2 - the spatial kernel;
B - the symbol of the banded part}
}
\value{
list of 3, the same as \code{Res}
}
\description{
Changes the separable-plus-banded model in list \code{Res} so that all three
kernels of the model are positive semi-definite.
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
Res <- spb(X,1)
Res_psd <- seppband_psd(Res)
}
