% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvband_nonstat.R
\name{shift_obs}
\alias{shift_obs}
\title{Shift an Observation to Calculate Off-diagonals of a Banded Covariance}
\usage{
shift_obs(X, i, j)
}
\arguments{
\item{X}{observation, a matrix}

\item{i}{row shift, has to be positive}

\item{j}{column shift, has to be negative (due to symmetry of the covariance,
it is enough to allow for negative row shift, forcing the column
shift to be positive, or vice versa)}
}
\value{
shifted observation adjunct by zeros to the size of \code{X}
}
\description{
Shift an Observation to Calculate Off-diagonals of a Banded Covariance
}
\examples{
( X <- matrix(9:1,3) )
shift_obs(X,1,2)
shift_obs(X,1,-2)
}
