% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adi.R
\name{solve_prec}
\alias{solve_prec}
\title{Solve the Preconditioning System}
\usage{
solve_prec(eigvals, x)
}
\arguments{
\item{eigvals}{eigenvalues of the preconditioner}

\item{x}{the right-hand side vector}
}
\value{
the solution as a matrix
}
\description{
Solves the linear system with the circulant preconditioner for the PCG, i.e.
it changes back the variables.
}
\examples{
eigvals <- matrix(9:1,3)
x <- stats::runif(9)
solve_prec(eigvals,x)
}
