% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spb.R
\name{spb}
\alias{spb}
\title{Separable-plus-banded Covariance Estimation}
\usage{
spb(
  X,
  d = NULL,
  stationary = TRUE,
  mu = NULL,
  predict = FALSE,
  maxd = NULL,
  mind = NULL,
  Folds = NULL,
  perc = c(1/4, 1/4)
)
}
\arguments{
\item{X}{data array of size \code{N} x \code{K1} x \code{K2}}

\item{d}{the bandwidth, an integer between \code{0} and \code{min(K1,K2)}}

\item{stationary}{a flag whether the banded part should be stationary,
defaults to \code{TRUE}}

\item{mu}{matrix of size \code{K1} x \code{K2}, the mean, defaults to the
empirical mean}

\item{predict}{flag whether prediction-based cross-validation should be used}

\item{maxd}{integer, maximum bandwidth considered by CV, defaults to
\code{min(K1,K2)/10}}

\item{mind}{minimum bandwidth considered by CV, defaults to zero}

\item{Folds}{number of folds for cross-validation, defaults to 10}

\item{perc}{vector of 2, if \code{predict=T}, what proportions (in time and
space) of every observation should be predicted. When
\code{perc}>=1, specific number of grid lines is predicted, e.g.
\code{perc}=1 corresponds to one-step-ahead prediction (in both
space and time)}
}
\value{
a list of objects, depending on \code{stationary} and \code{d}: \cr
\itemize{
\item \code{A1} - matrix of size \code{K1} x \code{K1}, the temporal kernel
\item \code{A2} - matrix of size \code{K2} x \code{K2}, the spatial kernel
\item \code{B} - matrix of size \code{d} x \code{d}, i.e. the symbol of
the banded part, if \code{stationary=T}. If \code{stationary=F}, then
the value depends on \code{d}: for \code{d=1} it is a matrix of size
\code{K1} x \code{K2} representing location-wise variances, while for
\code{d>1} a structure produced by \link{banded_nonstat_est} is returned
}

if cross-validation was conducted, the list contains also:
\itemize{
\item \code{cv} - the CV objective values: a larger value is better for fit-based
CV and a smaller value is better for the prediction-based CV (when
\code{predict=T})
\item \code{d} - the chosen bandwidth
}
}
\description{
Estimates components of \code{cov(X) = A1 x A2 + B} where \code{B} is banded
by \code{d}. If bandwidth \code{d} is not provided, it is estimated via
cross-validation. \cr
CV can be either fit-based, i.e tailored to minimize mean squared estimation
error, or prediction-based, i.e. tailored to minimize the mean squared
prediction error. The fit-based CV is faster. \cr
\code{B} is considered stationary by default, this can be
changed by toggling \code{stationary=F}.
}
\examples{
N <- 100
K1 <- K2 <- 7
X <- array(rnorm(N*K1*K2),c(N,K1,K2))
A1 <- array(0,c(K1,K1))
A1[,1] <- 1
A1[,2] <- -3:3
A1 <- A1 \%*\% t(A1)
A2 <- A1 <- mat_root(A1)
for(n in 1:N){
  X[n,,] <- A1 \%*\% X[n,,] \%*\% A2 + matrix(rnorm(K1*K2),K1)
}
# we have separable-plus-banded model, where B is stationary and d=1
Chat <- spb(X)
Chat$d # estimated bandwidth by fit-based CV
Chat <- spb(X,stationary=FALSE) # when we do not use stationarity
Chat$d # estimated bandwidth by fit-based CV without stationarity
}
