% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPT.R
\name{spt}
\alias{spt}
\title{Shifted Partial Tracing}
\usage{
spt(X, d = 0, standardize = TRUE)
}
\arguments{
\item{X}{data array of size \code{N} x \code{K1} x \code{K2}}

\item{d}{bandwidth, integer \code{>= 0}}

\item{standardize}{whether the results should be standardized,
defaults to True, option False is only for comparison
purposes with the partial inner product}
}
\value{
list of two:
\code{A1} - matrix of size \code{K1} x \code{K1}, the temporal kernel;
\code{A2} - matrix of size \code{K2} x \code{K2}, the spatial kernel;
}
\description{
Estimates the separable part of the separable-plus-banded model via Shifted
Partial Tracing. Choosing \code{d}=0 defaults to standard (non-shifted) partial
tracing estimator of the separable model.
}
\examples{
X <- array(runif(20*3*4),c(20,3,4))
spt(X,0)
}
