% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPT.R
\name{Ta}
\alias{Ta}
\title{Toeplitz Averaging}
\usage{
Ta(X, A1 = NULL, A2 = NULL, d)
}
\arguments{
\item{X}{data array of size \code{N} x \code{K1} x \code{K2}}

\item{A1, A2}{the separable part of the covariance}

\item{d}{bandwidth}
}
\value{
the stationary band - matrix of size \code{d} x \code{d}
}
\description{
Toeplitz Averaging
}
\examples{
X <- array(rnorm(5*10*7),c(5,10,7))
Ta(X,d=2)
}
