% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spb_goftest.R
\name{test_spb}
\alias{test_spb}
\title{Goodness-of-fit testing for separable-plus-banded model based on empirical
bootstrap.}
\usage{
test_spb(
  X,
  d = 1,
  stationary = TRUE,
  I = 1,
  J = 1,
  noB = 1000,
  mu = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{data array of size \code{N} x \code{K1} x \code{K2}}

\item{d}{the bandwidth, an integer between \code{0} and \code{min(K1,K2)}}

\item{stationary}{a flag whether the banded part should be stationary,
defaults to \code{TRUE}}

\item{I}{number of temporal eigenfunctions used to assess the fit}

\item{J}{number of spatial eigenfunctions used to assess the fit}

\item{noB}{number of bootstrap samples, defaults to 1000}

\item{mu}{matrix of size \code{K1} x \code{K2}, the mean, defaults to the
empirical mean}

\item{verbose}{flag whether the progress should be printed}
}
\value{
the p-value (a number - if small, it suggests the model is not valid)
}
\description{
Goodness-of-fit testing for separable-plus-banded model based on empirical
bootstrap.
}
\examples{
set.seed(123)
N <- 100
K1 <- K2 <- 7
X <- array(rnorm(N*K1*K2),c(N,K1,K2))
A1 <- array(0,c(K1,K1))
A1[,1] <- 1
A1[,2] <- -3:3
A1 <- A1 \%*\% t(A1)
A2 <- A1 <- mat_root(A1)
for(n in 1:N){
  X[n,,] <- A1 \%*\% X[n,,] \%*\% A2 + 2*matrix(rnorm(K1*K2),K1)
}
# the truth is separable+stationary (with d=1), i.e. it is also
# separable+diagonal at the same time
# set e.g. noB=1e4 in code below to get
test_spb(X,d=0,stationary=TRUE,noB=10)
# with d=0 the model is separable, which is rejected here
test_spb(X,d=1,stationary=TRUE,noB=10)
# separable+stationary model is not rejected
test_spb(X,d=1,stationary=FALSE,noB=10)
# separable+diagonal model is also not rejected
}
