% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_book_format}
\alias{to_book_format}
\title{Transform the symbol of B to format suitable for 2D FFT}
\usage{
to_book_format(band, K1, K2)
}
\arguments{
\item{band}{the symbol of B, matrix of size \code{d} x \code{d}, where
\code{d} is an integer}

\item{K1}{temporal grid size, integer \code{>= 2}}

\item{K2}{spatial grid size, integer \code{>= 2}}
}
\value{
matrix of size \code{2*K1} x \code{2*K2}
}
\description{
Transform the symbol of B to format suitable for 2D FFT
}
\examples{
to_book_format(matrix(runif(4),2), 3, 5)
}
